pyproject.toml
[build-system]
requires = ["setuptools>=68"]
build-backend = "setuptools.build_meta"

[project]
name = "obsidian-brain"
version = "0.1.0"
description = "Chat with your Obsidian vault locally using RAG and LLMs"
requires-python = ">=3.11"
readme = "README.md"
license = {text = "MIT"}
keywords = ["obsidian", "rag", "llm", "cli", "notes", "second-brain"]
dependencies = [
    "chromadb>=0.5.0",
    "ollama>=0.2.0",
    "watchdog>=4.0.0",
    "typer[all]>=0.12.0",
    "rich>=13.7.0",
    "python-frontmatter>=1.1.0",
    "tiktoken>=0.7.0",
    "python-dotenv>=1.0.0",
    "groq>=0.9.0",
    "google-generativeai>=0.8.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0",
    "pytest-mock>=3.12.0",
    "pytest-cov>=5.0.0",
]

[project.scripts]
brain = "brain.cli:app"

[tool.setuptools.packages.find]
where = ["."]
include = ["brain*"]

[tool.pytest.ini_options]
testpaths = ["tests"]
addopts = "-v --tb=short"
